Refactor lib folder
